changes to make postnote work similar to swiproc

+++ emu/port/dev.c
@@ -2,7 +2,7 @@
 #include	"fns.h"
 #include	"error.h"
 
-extern ulong	kerndate;
+extern u32	kerndate;
 
 void
 mkqid(Qid *q, vlong path, ulong vers, int type)
